4th Semester/Programming for Artificial Intellegence (Lab)/Project/requirements.txt
google-genai
pandas
numpy
sentence-transformers
faiss-cpu
python-dotenv